funnel_automation/requirements.txt
streamlit==1.31.0
pandas==2.2.0
requests==2.32.0
python-dotenv==1.0.0
google-analytics-data==0.18.0
google-api-python-client==2.116.0
google-auth==2.27.0
google-cloud-bigquery==3.25.0
plotly==5.18.0
